backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
SQLAlchemy==2.0.21

# Computer Vision and ML
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-image==0.21.0

# OCR Libraries
pytesseract==0.3.10
easyocr==1.7.0

# Face Recognition
face-recognition==1.3.0
dlib==19.24.2
deepface==0.0.79

# Image Processing
mediapipe==0.10.7
imutils==0.5.4

# Blockchain Integration (for future implementation)
web3==6.11.1
py-solc-x==1.12.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30
python-dateutil==2.8.2

# Development
pytest==7.4.2
pytest-flask==1.2.0
